requirements.txt
PyQt5
pygost
openpyxl
pycryptodome
